Implement modifier N (closes #9); Fix lots of integer truncation warnings
git-svn-id: svn+ssh://beta.visl.sdu.dk/usr/local/svn/repos/visl/tools/vislcg3/trunk@12903 cb2587b3-b7ff-0310-8c81-a2a651690ada

+++ src/AST.hpp
@@ -92,7 +92,7 @@ const char* ASTType_str[NUM_ASTTypes] = {};
 struct ASTHelper;
 struct ASTNode {
 	ASTType type;
-	uint32_t line;
+	size_t line;
 	const UChar *b, *e;
 	std::vector<ASTNode> cs;
