+++ src/GrammarApplicator_runGrammar.cpp
@@ -122,7 +122,7 @@ void GrammarApplicator::runGrammarOnText(std::istream& input, std::ostream& outp
 		++lines;
 		size_t offset = 0, packoff = 0;
 		// Read as much of the next line as will fit in the current buffer
-		while (u_fgets(&line[offset], line.size() - offset - 1, input)) {
+		while (u_fgets(&line[offset], static_cast<int32_t>(line.size() - offset - 1), input)) {
 			// Copy the segment just read to cleaned
 			for (size_t i = offset; i < line.size(); ++i) {
 				// Only copy one space character, regardless of how many are in input
